Update otel-collector-custom sample
As of Config Sync v1.17.0.

This is a sample of custom otel-collector configuration that only
contains Prometheus pipeline, and the Monarch pipeline that is designed
for internal operational monitoring.

To apply, run `kubectl apply -f otel-collector-custom.yaml` on cluster.
To revert to original setting, remove with `kubectl delete -f
otel-collector-custom.yaml`.